+++ CHIP-8 Emulator/src/Chip8.h
@@ -157,7 +157,7 @@ private:
 	uint16_t opcode;
 
 	std::default_random_engine randGen;
-	std::uniform_int_distribution<uint8_t> randByte;
+	std::uniform_int_distribution<unsigned int> randByte;
 
 	typedef void (Chip8::* Chip8Func)();
 	Chip8Func table[0xF + 1]{ &Chip8::OP_NULL };
